requirements.txt
Flask>=3.1.2
python-telegram-bot>=22.5
gunicorn>=23.0.0
waitress==3.0.0
httpx>=0.28.0
APScheduler>=3.10.4
psutil>=5.9.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
APScheduler
Flask
gunicorn
httpx
psutil
psycopg2-binary
pytest
pytest-asyncio
pytest-cov
pytest-mock
python-dotenv
python-telegram-bot
waitress
python-docx
